ebuild: add the get_real_uri() method that returns the resolved URI.
GIT enables substitutions in URIs using the uri.<base>.insteadOf configuration
variable. The get_uri() still returns the original URI. But get_real_uri()
returns the URI with all substitutions applied. Internally, the URI was already
resolved in order to check if it was a valid GIT URI. But it was not stored nor
returned by any existing method.