ebuild: enable override of revision with an empty value.
We previously used this pattern to define revisions:
: ${SCM_REVISION:="xxx"}

Unfortunately, it does not enable us to change the revision to an empty value.
It can be useful when one wants to test an ebuild against the latest revision
of a branch. We can get the expected result using this pattern instead:
: ${SCM_REVISION="xxx"}